+++ src/api/index.ts
@@ -1,0 +1,3 @@
+export * from './api-platform.enum';
+export * from './api-version.enum';
+export * from './swagger-version.enum';