+++ src/core/domain/entities/user.entity.ts
@@ -1,0 +1,23 @@
+import { Types } from 'mongoose';
+import { IUser } from '../interfaces';
+
+export class UserEntity implements IUser {
+  _id?: Types.ObjectId;
+  email: string;
+  password?: string | null;
+  name: string;
+  provider?: 'local' | 'google';
+  providerId?: string;
+  avatar?: string;
+  roles?: string[];
+  createdAt?: Date;
+  updatedAt?: Date;
+
+  constructor(partial: Partial<IUser>) {
+    Object.assign(this, partial);
+  }
+
+  hasRoles(role: string): boolean {
+    return this.roles?.includes(role) ?? false;
+  }
+}