+++ test/e2e/constants/index.ts
@@ -1,0 +1,2 @@
+export * from './endpoint-topics.constant';
+export * from './endpoint-subtopics.constant';